% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_create_grille_poly.R
\name{create_grille_poly}
\alias{create_grille_poly}
\title{create_grille_poly}
\usage{
create_grille_poly(
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  n_ligne = NULL,
  n_col = NULL,
  res = NULL,
  poly = NULL,
  buffer = 0,
  type_grille = "square"
)
}
\arguments{
\item{xmin}{as numeric}

\item{xmax}{as numeric}

\item{ymin}{as numeric}

\item{ymax}{as numeric}

\item{n_ligne}{as integer}

\item{n_col}{as integer}

\item{res}{as integer : resolution dela grille (dans l'unite de travail)}

\item{poly}{a shapefile, sf format poly<-sf::st_read(fich_poly,quiet=TRUE)}

\item{buffer}{l'extension a rajouter autour des valeurs limites ou du shape donnees (dans l'unite de travail)}

\item{type_grille}{as text : square ou hexagonal}
}
\description{
permet de creer une grille reguliere. Plusieures possibilite : soit on rentre les xmin/xmax ymin/ymax, soit on rentre un polygone (sp) dasn poly,
soit on laisse les 3 parametres nul et alors la fonction va ouvrir une boite de dialogue pour demander de choisir un shapefile.
Pour la resolution de la grille : soit on rentre nligne et ncol, soit on rentre la resolution. Si les trois parametres sont rentres, la fonction ne prendra que la resolution.
On peut rajouter un buffer, dans ce cas un contor sera pris d'autant autour du shapefile ou des coordonees xy
}
\examples{
\dontrun{
g<-create_grille_poly(xmin=368880,xmax=410068,ymin=6554456,ymax=6582756,n_ligne=100,n_col=100)
g<-create_grille_poly(poly=fr,res=1,type_grille="HEXAGONAL") # fr est un shapefile au format sf
plot(g)
g2<-create_grille_poly(poly=fr,nligne=10,ncol=10)
g2<-create_grille_poly(poly=fr,res=1,buffer=2)
}
}
