% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame2graphiqueanigif.R
\name{data_frame2graphiqueanigif}
\alias{data_frame2graphiqueanigif}
\title{data_frame2graphiqueanigif}
\usage{
data_frame2graphiqueanigif(
  df,
  vitesse = 10,
  outdos = NULL,
  titre = "",
  xylabs = NULL,
  vide_temp_avant = TRUE,
  size = c(700, 300)
)
}
\arguments{
\item{df}{as dataframe au format 3 colonnes : x, y, groupe . x est soit numerique, soit date. Seule la position du champs compte, pas son nom.}

\item{vitesse}{as  float : speed of the anigif}

\item{outdos}{dossier de sortie}

\item{titre}{title to write above the anigif}

\item{xylabs}{as a vector of two string for respectively xlab and ylab. # titre des axes x et y, si non renseignes dasn l'appel de fonction, ce sera le titre des colonnes}

\item{vide_temp_avant}{delete all files in temp dir before.   vide le dossier si il existe deja et si vide_temp_avant=T}

\item{size}{a vector of 2 integer for the out size of the anigif}
}
\description{
Permet de creer un graphique anime où la courbe des valeurs se trace au fur et a mesure. Attention, ne fonctionne que si ImageMagick (logiciel) est installe sur le poste.
}
\examples{
df<-data.frame(x=1:20,y=rnorm(20),groupe=rep(c("a","b"),10))
data_frame2graphiqueanigif(df,outdos="C:/TEMP/",vide_temp_avant=TRUE)
}
