% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame2map.R
\name{data_frame2map}
\alias{data_frame2map}
\title{data_frame2map}
\usage{
data_frame2map(
  data,
  crs_epsg = 4326,
  champs = NA,
  nom_champsx = "x",
  nom_champsy = "y",
  nom_champs_etiquette = NULL,
  type_carte = "Classique",
  taille_etiquettes = 0.8,
  nom_champs_popup = NULL
)
}
\arguments{
\item{data}{as dataframe avec à minima une colonne pour les x, une colonne pour les y et une colonne pour le champs de valeur}

\item{crs_epsg}{as text}

\item{champs}{as text le nom du champs de valeur pour le plot? si champs=NA : seule l'emplacement des points sera dessiné}

\item{nom_champsx}{as text}

\item{nom_champsy}{as text}

\item{nom_champs_etiquette}{as text pour affichage des etiquettes}

\item{type_carte}{as text : "Leaflet" ou "Classique". Tout autre nom que Leaflet donnera une carte classique (defaut).}

\item{taille_etiquettes}{as real : taille d'affichage des etiquettes en mode classique uniquement}

\item{nom_champs_popup}{as text pour affichage quand on click sur le point}
}
\description{
Permet de visualiser sur une carte un data frame contenant des données spatiales. Il corrige un bug connu de décallage des données
représentées sur les fonds Google map. Pour ne pas représenter un champs de valeur mais juste la position des points : champs=NA
}
\note{
auteur : AH/AtmoNA, modification : 22/04/2024 : ajout de la possibilite de creer une carte avec Leaflet
}
\examples{
\dontrun{
# tracer une diagonale de point au dessus de l'ouest de l'afrique associés à des valeurs aléatoires
df<-data.frame(x=-5:4,y=41:50,val=rnorm(400))
data_frame2map(df,champs="val")
data_frame2map(df,champs="val",nom_champs_etiquette="x",taille_etiquettes=5)
data_frame2map(df,champs="val",nom_champs_etiquette=NULL,type_carte="Leaflet")
}
}
