% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodair.R
\name{data_geodair}
\alias{data_geodair}
\title{data_geodair}
\usage{
data_geodair(
  polluant,
  date_d,
  date_f,
  stat = "MoyH",
  codesite = NULL,
  aasqa = NULL
)
}
\arguments{
\item{polluant}{as text :selon les codes iso de geodair ("03" : NO2, "10": CO, "24" : PM10,"T1":Nb PM 7nm-1µm, "T3" : Nb PM 10nm-1µm...). un fichier contenant les codes iso est sur le reseau atmona partenariat/outils/geodair}

\item{date_d}{as text : date de debut au format YYYY-MM-DD}

\item{date_f}{as text : date de debut au format YYYY-MM-DD}

\item{stat}{as text : "MoyH" : moyenne horaire, "MoyA": myenne annuelle , "MoyJ" moyenne jour,"MaxJH" max horaire journalier}

\item{codesite}{as text : le code du site a conserver au format GEODAIR}

\item{aasqa}{as text "ATMO SUD", "ATMO OCCITANIE" "ATMO NOUVELLE-AQUITAINE" "ATMO AUVERGNE-RHÔNE-ALPES". Pour toute la France : aasqa=NULL}
}
\description{
: la fonction permet de recuperer plusieurs jours de donnees pour un polluant a partir de GEODAIR publique.Elle utilise la fonction get_data_pub() qui fait fonctionner l'API pour un jour un polluant.
L'API telecharge les valeurs pour tous les sites de France. Les parametres "codesite" et "aasqa" permettent de filtrer le resultat telecharge, mais toutes les donnees
du territoire seront quand meme telechargees (pas le choix). Attention : le nombre de requete (jour) par heure calendaire est limite (ex : 3000 jours /heure)
}
\examples{
\dontrun{
library(AnisTools)
vals<-data_geodair(polluant="T3",date_d="2024-11-01",date_f="2024-12-17",aasqa=NULL)
vals<-NULL
vals<-data_geodair(polluant="T3",date_d="2023-01-01",date_f="2023-01-02")
}
}
