% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_1var.R
\name{data_summary_1var}
\alias{data_summary_1var}
\title{data_summary_1var}
\usage{
data_summary_1var(data, varname, groupnames)
}
\arguments{
\item{data}{dataframe}

\item{varname}{as character}

\item{groupnames}{as character}
}
\description{
data_summary_1var
}
\examples{
df<-data.frame(x=1:20,y=rnorm(20),groupe=rep(c("a","b","c","d"),5))
data_summary_1var(df,"y","groupe")
}
