% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_2var.R
\name{data_summary_2var}
\alias{data_summary_2var}
\title{data_summary_2var
return(res)}
\usage{
data_summary_2var(data, varname1, varname2, groupnames)
}
\arguments{
\item{data}{un data frame, avec 2 colonnes qui contiennent les valeurs à comparer, ainsi qu'un nom de champs pour grouper les comparaisons}

\item{varname1}{as character : nom du premier champs}

\item{varname2}{as character: nom du second champs}

\item{groupnames}{as character : nom du champs qui comporte le groupe.
Pour ne pas utiliser de groupes : groupnames=NULL}
}
\description{
sert à carculer des stats sur des séries de valeurs
data_summary_2 : format : un data frame, avec 2 colonnes qui contiennent les valeurs à comparer,
et une autre colonne qui contient le nom des groupes (ex : mois, ...)
}
\examples{
df<-data.frame(x=1:20,y1=rnorm(20),y2=rnorm(20)*10,groupe=rep(c("a","b","c","d"),5))
data_summary_2var(df,varname1="y1",varname2="y2",groupnames="groupe")
#library(openair) # aussi possible de cette maniere
#modStats(df,"y1","y2")
}
