% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_stat.R
\name{data_summary_stat}
\alias{data_summary_stat}
\title{data_summary_stat}
\usage{
data_summary_stat(data, varname, groupnames)
}
\arguments{
\item{data}{as dataframe. 3 colonnes : une colonne d'identifiant qui s'appelle obligatoirement "ID". Une colonne de valeurs, une colonne avec les groupes.}

\item{varname}{le nom de la variable (du champs) du data frame}

\item{groupnames}{le nom du champs du data frame qui contient les groupes}
}
\value{
dataframe
}
\description{
: permet de dire quel ID correspond au min et au max d'une variable donnee dans chaque groupe.
}
\examples{
df<-data.frame(ID=1:20,y=rnorm(20),groupe=rep(c("a","b","c","d"),5))
data_summary_stat(df,"y","groupe")
}
