% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrader_couleurs.R
\name{degrader_couleurs}
\alias{degrader_couleurs}
\title{degrader_couleurs}
\usage{
degrader_couleurs(coul, seuilmini, seuilmaxi, range = 4)
}
\arguments{
\item{coul}{colors to interpolate; must be a valid argument to col2rgb(). par exemple : c("red", "green")}

\item{seuilmini}{: minima de l'échelle}

\item{seuilmaxi}{: maxima de l'échelle}

\item{range}{: nombre de classes}
}
\description{
: permet de creer une echelle de couleur en dégradé avec le nombre de classes demandées a partie des couleurs choisies.
}
\examples{
degrader_couleurs(c("red", "green"),0,10,4)
degrader_couleurs(coul=c("red", "green","purple"),0,10,4)
}
