% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_package.R
\name{detach_package}
\alias{detach_package}
\title{detach_package}
\usage{
detach_package(pkg, character.only = FALSE)
}
\arguments{
\item{pkg}{nom du package}

\item{character.only}{}
}
\description{
permet de detacher un package proprement et sans erreurs
}
\examples{
\dontrun{
detach_package("vegan", TRUE)
}
}
