% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodair.R
\encoding{UTF-8}
\name{get_data_pub}
\alias{get_data_pub}
\title{Connexion a l'API publique de GEOD'AIR pour le telechargement des donnees}
\usage{
get_data_pub(
  polluant,
  date,
  stat = "MoyH",
  user_pub = NULL,
  pwd_pub = NULL,
  api_key_pub = NULL
)
}
\arguments{
\item{polluant}{as text :selon les codes iso de geodair ("03" : NO2, "10": CO, "24" : PM10,...)}

\item{stat}{as text : "MoyH" : moyenne horaire, "MoyA": myenne annuelle , "MoyJ" moyenne jour,"MaxJH" max horaire journalier}

\item{user_pub}{: id (et pas mail) pour l'espace public de geodair. (Il semblerait que le mauvais nom de user n'impacte pas la requete).}

\item{pwd_pub}{: pwd pour l'espace public de geodair}

\item{api_key_pub}{: l'api de lespace public, qui est cree une fois pour toute ici : https://www.geodair.fr/donnees/api}

\item{date_d}{as text : date de debut au format YYYY-MM-DD}

\item{date_f}{as text : date de debut au format YYYY-MM-DD}
}
\value{
un objet de classe S3 geodair
}
\description{
fait fonctionner l'API GEODAIR publique. la fonction ne permet de telecharger qu'une seule journee. Pour plusieures journées, utiliser data_geodair()
}
\examples{
\dontrun{
data<-NULL
data<-get_data_pub(polluant="24",date="2024-11-23")
}

}
