% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_geofla.R
\name{get_last_geofla}
\alias{get_last_geofla}
\title{get_last_geofla}
\usage{
get_last_geofla(echelle, dos_sortie = NULL)
}
\arguments{
\item{echelle}{}
}
\description{
la fonction permet de télécharger les couches sig  administratives cgeofla. Certains couches sont téléchargées individuellement
en lamb93 (commune, departement,canton arrondissement) puis converties en wgs84. Les autres proviennent du pack complet, où seules les couches
souhaitées sont extraites en wgs84. Si "tout" est demandé, c'est le pack
complet en wgs84 qui est extrait (avec plein de couches inutiles).
}
\examples{
\dontrun{
get_last_geofla(echelle="region")
get_last_geofla(echelle="commune",dos_sortie="C://TEMP")
}
}
