% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dataframe.R
\name{project_dataframe}
\alias{project_dataframe}
\title{project_dataframe}
\usage{
project_dataframe(d, srid_in = NULL, srid_out = NULL, nomx = NULL, nomy = NULL)
}
\arguments{
\item{d}{: dataframe with 2 columns minimum : "x", "y" ou "lat","lon"}

\item{srid_in}{: srid de la projection source : https://app.dogeo.fr/Projection/#/point-to-coords}

\item{srid_out}{: srid de la projection de sortie : https://app.dogeo.fr/Projection/#/point-to-coords}

\item{nomx}{as text : le nom du champs contenant les x dans le dataframe d'entree}

\item{nomy}{as text : le nom du champs contenant les y dans le dataframe d'entree}
}
\description{
: permet de convertir les coordonees geographiques d'un data frame d'un système vers un autre. On precise le nom des colonnes x et y
}
\examples{
d<-data.frame(id=1:5,lon=seq(-3,3,length.out=5),lat=seq(45,46,length.out=5))
}
