% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_et_hillshade2map.R
\name{raster_et_hillshade2map}
\alias{raster_et_hillshade2map}
\title{raster_et_hillshade2map}
\usage{
raster_et_hillshade2map(r, zoom_elev = 5)
}
\arguments{
\item{r}{un raster avec son crs défini}
}
\description{
raster_et_hillshade2map
}
\examples{
\dontrun{
library(raster)
fich<-"N:/4_developpements/COPERNICUS/IASI/NH3/nh3_2020-03-10_2020-04-20EU_mg_cm-2_lisse.tif"
raster(fich)->r
r <- crop(r, extent( -9, 9, 40, 50))
}

}
