% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_point2map.R
\name{sf_point2map}
\alias{sf_point2map}
\title{sf_point2map}
\usage{
sf_point2map(
  data,
  champ = NULL,
  zoom_ = 3,
  titre = NULL,
  etiquettes = NULL,
  representation = "taille",
  placement_etiquette = "droite"
)
}
\arguments{
\item{data}{: point shape, format sp ou sf}

\item{champ}{: as text nom du champ à représenter}

\item{zoom_}{: zoom pour la carte du fond. 6 par defaut. (1 : monde, + de 10 : ville)}

\item{titre}{as text : le titre à ajouter en haut de la carte (NULL par defaut = blanc)}

\item{etiquettes}{as text : le champs à utiliser pour représenter des étiquettes à droite des points ou sur le point (cf placement_etiquettes)}

\item{representation}{as text : "taille" ou "couleur" pour la représentation du champs des points}

\item{placement_etiquettes}{as text : "droite" ou "dessus"}
}
\value{
une carte au format ggmap
}
\description{
permet de créer une carte avec un fond OSM pour représenter une couche de points sf ou sp. On peut aussi (optionnel) afficher un champs en etiquette
}
\examples{
\dontrun{
data_keep2<-data.frame(xco=c(-1,0,0.5),yco=c(46,46.2,46.3),desc=c("g","f","p"),desc2=1:3)
data_keep2<-st_as_sf(x=data_keep2,coords = c(1, 2), remove = FALSE, crs = st_crs("+proj=longlat +ellps=WGS84"))
sf_point2map(data_keep2,"desc2",zoom_=9,titre="localisation des mesures",etiquette="desc",representation="couleur",placement_etiquette="droite")
}
}
