% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_polygon2map.R
\name{sf_polygon2map}
\alias{sf_polygon2map}
\title{sf_polygon2map}
\usage{
sf_polygon2map(
  sf_poly = NULL,
  champs = NULL,
  titre = NULL,
  nom_legende = NULL,
  echelle_couleur_viridis = "viridis",
  interval_leg = NULL,
  transparence = 0.4,
  contour_poly = "oui",
  legende = TRUE,
  zoom_fond = 10,
  rev_col = 1,
  echelle_cut = NULL,
  n_class = 5,
  cex_text = 1,
  etiquettes = NULL,
  type_carte = "stamen_terrain"
)
}
\arguments{
\item{sf_poly}{un vecteur polygone de type sf. Si non renseigne, la foncion ouvre une fenetre pour selectionner le fichier.}

\item{champs}{as text un champs numerique ou nom pour la representation}

\item{titre}{as text titre a ajouter au graphique}

\item{nom_legende}{as text titre de la legnde. Si NULL, sera le nom de la variable}

\item{echelle_couleur_viridis}{as text. "viridis" "magma" "plasma" "inferno" "cividis" "mako" "rocket" "turbo" . cf https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html}

\item{interval_leg}{as vector of 2 num : ex : c(0,3) : impose un interval lineaire entre les 2 bornes. Les bornes peuvent depasser ou être inferieures à la gamme des valeurs du champs represente.}

\item{transparence}{transparence de la couche polygone}

\item{contour_poly}{as text : faire apparaitre le contour ou non pour les polygones ("oui"/"non"), avec la meme couleur que le remplissage ("oui_idem")}

\item{legende}{as boolean : est ce que la legende doit apparaitre}

\item{zoom_fond}{as interger : map zoom, an integer from 3 (continent) to 21 (building), default value 10 (city). openstreetmaps limits a zoom of 18, and the limit on stamen maps depends on the maptype. "auto" automatically determines the zoom for bounding box specifications, and is defaulted to 10 with center/zoom specifications. maps of the whole world currently not supported.}

\item{rev_col}{as interger : 1 ou -1 . -1 : va du plus clair au plus fonce.1: l'inverse.}

\item{echelle_cut}{as text.Correspond au type d'echelle a utiliser, NULL par defaut. valeurs possibles : one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks", "dpih", "headtails", or "maximum"}

\item{n_class}{as integer : imposer le nombre de classe. 5 par defaut}

\item{cex_text}{as numeric : facteur multiplicatif proportionel qui sera applique a la taille du texte du titre, de la legende et du titre de la legende. 1 par defaut}

\item{type_carte}{options available  "stamen_terrain", "stamen_toner", "stamen_toner_lite", "stamen_watercolor", "stamen_terrain_background", "stamen_toner_background", "stamen_terrain_lines", "stamen_terrain_labels", "stamen_toner_lines", "stamen_toner_labels" (Stadia Maps)}
}
\description{
sf_polygon2map
}
\details{
pour utiliser la fonction avec un raster r3 : st_as_sf(rasterToPolygons(r3 ))->poly
}
\note{
dernieres modifications : septembre 2023 : ajout de la possibilite de mettre des etiquettes
!! octobre 2023 : stamen nous laisse tomber pour les fonds de carte. Desormais il faut demander une cle API a stadiamaps. ggmap doit etre reinstalle.
https://docs.stadiamaps.com/guides/migrating-from-stamen-map-tiles/
remove.packages("ggmap")
devtools::install_github("stadiamaps/ggmap")
library(ggmap)
register_stadiamaps("xxx-xxx-xxx-xxx",write = TRUE). write=true pour que ce soit permanent et qu'on ai plus a faire de register pour ce poste
us <- c(left = -125, bottom = 25.75, right = -67, top = 49)
get_stadiamap(us, zoom = 5, maptype = "stamen_toner") \%>\% ggmap()
}
\examples{
\dontrun{
fich_com_na<-"K:\\\\Partenariat_Innovation\\\\SIG\\\\France\\\\GEOFLA\\\\communes_NA.shp"
com_na<-sf::st_read(fich_com_na,quiet=TRUE)
sf_polygon2map(sf_poly=com_na,champs="iasi2",zoom_fond=7,contour_poly = "non",interval_leg=c(1000,4000),echelle_cut="fixed",n_class=40)

sf_polygon2map(titre="Titre du graphique", nom_legende="variable cartographiee")
}
}
